insight-engine/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
python-dotenv==1.0.0
praw==7.7.1
textblob==0.17.1
nltk==3.8.1
beautifulsoup4==4.12.2
requests==2.31.0 